Update hyphenation files
This commit adds all hyphenation files from LibreOffice as located in
https://github.com/LibreOffice/dictionaries.

Additionally a script is added to the tools-directory to download the
ZIP-file of the LibreOffice dictionaries and update the dictionaries
based on any new release. This should ease retrieval of new or updated
hyphenation-dictionaries.

The script can be invoked via
`./tools/updateHyphenationFilesFromLibreOffice.sh` and hanles everything
else from there. The files need to be added to git manually afterwards.